restcountries_tidy_data

A dataset containing tidied information about countries from the Restcountries API.

Usage

restcountries_tidy_data

Format

A data frame with several rows and the following columns:

tld

Top-level domain(s) associated with the country.

common_name

Common name of the country.

official_name

Official name of the country.

cca2

Country code (2-letter).

cca3

Country code (3-letter).

fifa

FIFA code of the country.

independent

Independence status (TRUE/FALSE).

status

Country status (e.g., officially assigned).

un_member

Whether the country is a UN member (TRUE/FALSE).

region

Geographic region.

subregion

Subregion.

population

Population of the country.

capital

Capital city of the country.

capital_lat

Latitude of the capital city.

capital_lon

Longitude of the capital city.

continents

Continent(s) the country is part of.

lat

Latitude of the country.

lon

Longitude of the country.

landlocked

Whether the country is landlocked (TRUE/FALSE).

borders

Countries that share a border.

area

Total area of the country in square kilometers.

start_of_week

Day the week starts (e.g., Monday).

timezones

Timezones applicable to the country.

root

Root of the country calling code.

suffixes

Suffixes of the country calling code.

car_side

Which side of the road cars drive on.

googlemaps

Google Maps link for the country.

openstreetmaps

OpenStreetMap link for the country.

flags_png

URL to PNG image of the country flag.

flags_svg

URL to SVG image of the country flag.

flags_alt

Alternative text for the country flag.

currencies

Currencies used in the country.

languages

Languages spoken in the country.

currency_name

Name of the primary currency used.

currency_symbol

Symbol of the primary currency used.

calling_code

Calling code(s) associated with the country.

Source

Data obtained from the Restcountries Json file and processed for use in this package.

Details

This dataset includes a variety of country-level data such as country codes, names, capitals, regions, subregions, continents, currencies, population, geographic coordinates, languages, and more.

Examples

# Load the dataset and view the first few rows
data(restcountries_tidy_data)
head(restcountries_tidy_data)
#> # A tibble: 6 × 36
#>   tld   common_name official_name cca2  cca3  fifa  independent status un_member
#>   <chr> <chr>       <chr>         <chr> <chr> <chr> <lgl>       <chr>  <lgl>    
#> 1 .aw   Aruba       Aruba         AW    ABW   "ARU" FALSE       offic… FALSE    
#> 2 .af   Afghanistan Islamic Repu… AF    AFG   "AFG" TRUE        offic… TRUE     
#> 3 .ao   Angola      Republic of … AO    AGO   "ANG" TRUE        offic… TRUE     
#> 4 .ai   Anguilla    Anguilla      AI    AIA   "AIA" FALSE       offic… FALSE    
#> 5 .ax   Åland Isla… Åland Islands AX    ALA   ""    FALSE       offic… FALSE    
#> 6 .al   Albania     Republic of … AL    ALB   "ALB" TRUE        offic… TRUE     
#> # ℹ 27 more variables: region <chr>, subregion <chr>, population <int>,
#> #   capital <chr>, capital_lat <dbl>, capital_lon <dbl>, continents <chr>,
#> #   lat <dbl>, lon <dbl>, landlocked <lgl>, borders <list>, area <dbl>,
#> #   start_of_week <chr>, timezones <chr>, root <chr>, suffixes <chr>,
#> #   car_side <chr>, googlemaps <chr>, openstreetmaps <chr>, flags_png <chr>,
#> #   flags_svg <chr>, flags_alt <chr>, currencies <chr>, languages <chr>,
#> #   currency_name <chr>, currency_symbol <chr>, calling_code <chr>